divide JetsonGPIO.h into multiple headers
