+++ src/Callback.cpp
@@ -23,7 +23,7 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 DEALINGS IN THE SOFTWARE.
 */
 
-#include "JetsonGPIO.h"
+#include "JetsonGPIO/Callback.h"
 
 namespace GPIO
 {
